Make installed wheels reproducible
Pip installed wheels are not reproducible currently. The direct_url
files encode an installation path and the installed wheels compile
the python files at their location, not their final install location
which is incorrect.

To fix this, simply disable the direct_urls and pass the "root" to
the python compile function to strip that path out of the compiled
files.

A version of this patch, perhaps stripping root from the direct_urls
may be something that could be considered by upstream.